refactor(jwt): JwtConstants를 object로 변경
JwtConstants를 class + companion object에서 object로 변경하여
코드를 단순화하고 Kotlin 관례를 따르도록 수정했습니다.

변경 사항:
- JwtConstants: class + companion object → object
- CookieHelper: JwtConstants.Companion 제거
- AuthApiController: import 경로 수정
