+++ main-server/src/main/kotlin/com/techtaurant/mainserver/security/infrastructure/in/AuthApiController.kt
@@ -2,7 +2,7 @@ package com.techtaurant.mainserver.security.infrastructure.`in`
 
 import com.techtaurant.mainserver.common.dto.ApiResponse
 import com.techtaurant.mainserver.security.aop.AuthRestController
-import com.techtaurant.mainserver.security.cookie.CookieHelper
+import com.techtaurant.mainserver.security.helper.CookieHelper
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import jakarta.servlet.http.HttpServletResponse
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 class AuthApiController(
     private val cookieHelper: CookieHelper,
 
-) {
+    ) {
     @Operation(summary = "로그아웃", description = "쿠키에서 토큰을 삭제하여 로그아웃합니다")
     @ApiResponses(
         value = [
