+++ main-server/src/main/kotlin/com/techtaurant/mainserver/security/jwt/JwtConstants.kt
@@ -1,8 +1,6 @@
 package com.techtaurant.mainserver.security.jwt
 
-class JwtConstants {
-    companion object {
-        const val ACCESS_TOKEN_COOKIE = "accessToken"
-        const val REFRESH_TOKEN_COOKIE = "refreshToken"
-    }
+object JwtConstants {
+    const val ACCESS_TOKEN_COOKIE = "accessToken"
+    const val REFRESH_TOKEN_COOKIE = "refreshToken"
 }
